pyproject.toml
[project]
name = "dumbvlms"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "matplotlib>=3.10.1",
    "numpy>=2.2.4",
    "pillow>=11.1.0",
    "pyyaml>=6.0.2",
]
